Trabalhando com Mongodb e Sprint Boot, arrumando os testes e trocando o tipo da variável id do usuario
