+++ README.md
@@ -1,2 +1,10 @@
 # spring-boot-mockito
 How to write Unit test case using mockito framework 
+
+#Atualização para evitar erro de duplicação de chave do mongodb#
+- mudamos o tipo do id do usuário para String para evitar duplicação de chave
+- mudamos o teste do controler do usuário para a pasta específica, desacoplando da pasta da aplicação
+- testamos a api no Postaman com sucesso
+
+#Geração de endpoint#
+- foi gerado o endpoint caso necessite testar
