+++ src/main/java/com/javatechie/spring/mockito/api/model/User.java
@@ -1,11 +1,18 @@
 package com.javatechie.spring.mockito.api.model;
 
+import org.bson.types.ObjectId;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.data.annotation.Id;
+import org.springframework.data.mongodb.core.index.Indexed;
 import org.springframework.data.mongodb.core.mapping.Document;
+
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
 import lombok.NoArgsConstructor;
+import lombok.Setter;
 import lombok.ToString;
 
 @Document(collection = "users")
@@ -13,9 +20,12 @@ import lombok.ToString;
 @AllArgsConstructor
 @NoArgsConstructor
 @ToString
+@EqualsAndHashCode(onlyExplicitlyIncluded = true)
+@Getter
+@Setter
 public class User {
 	@Id
-	private int id;
+	private  String id;
 	private String name;
 	private int age;
 	private String address;
